Creating Project-1
The data that is used in this project was collected and made available by Dr. Kristen Gorman and Palmer Station, Antartica LTER, a member of the 'Long Term Ecological Research Network'.
In this project, a detailed  exploratory data analysis is done on the data first, then the data was split into two split ratios which were, 60-20-20, and 80-10-10. Following that, optimal hyperparameters for Logistic Regression, Multilayer Percepton and Random Forest for both split ratios is analyzed.

After that, for each algorithm the learning curve is plotted on both split ratios. The plot illustrated  the training score and cross validation score over the number of training examples.
Furthermore, performance of each algorithm on both split ratios is found and compared to each other.
In conclusion, the final analysis was consisted of finding best split ratio and algorithm.